split up index into components
