+++ website/src/components/Feature/styles.module.scss
@@ -1,0 +1,10 @@
+.features {
+  align-items: center;
+  padding: 2rem 0;
+  width: 100%;
+}
+
+.featureImage {
+  height: 200px;
+  width: 200px;
+}
