+++ website/src/pages/index.tsx
@@ -1,89 +1,17 @@
 import React from "react";
-import classnames from "classnames";
-import Layout from "@theme/Layout";
 import Link from "@docusaurus/Link";
-import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import useBaseUrl from "@docusaurus/useBaseUrl";
-import styles from "./styles.module.scss";
 import Translate, { translate } from "@docusaurus/Translate";
+import useBaseUrl from "@docusaurus/useBaseUrl";
+import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+import Layout from "@theme/Layout";
+import classnames from "classnames";
+import { Feature } from "../components/Feature/index";
+import { Preview } from "../components/Preview";
+import { features } from "../data/features";
+import styles from "./styles.module.scss";
 
-const features = [
-  {
-    title: <Translate id="homepage.compile_safe_title">Compile safe</Translate>,
-    imageUrl: "img/undraw_security.svg",
-    description: (
-      <Translate
-        id="homepage.compile_safe_body"
-        values={{ ProviderNotFound: <code>ProviderNotFoundException</code> }}
-      >
-        {`No more {ProviderNotFound} or forgetting to handle loading
-        states. Using Riverpod, if your code compiles, it works.`}
-      </Translate>
-    ),
-  },
-  {
-    title: (
-      <Translate id="homepage.unlimited_provider_title">
-        Provider, without its limitations
-      </Translate>
-    ),
-    imageUrl: "img/undraw_friendship.svg",
-    description: (
-      <Translate id="homepage.unlimited_provider_body">
-        Riverpod is inspired by Provider but solves some of it's key issues such as
-        supporting multiple providers of the same type; awaiting asynchronous
-        providers; adding providers from anywhere, ...
-      </Translate>
-    ),
-  },
-  {
-    title: (
-      <Translate id="homepage.no_flutter_dependency_title">
-        Doesn't depend on Flutter
-      </Translate>
-    ),
-    imageUrl: "img/undraw_programming.svg",
-    description: (
-      <Translate
-        id="homepage.no_flutter_dependency_body"
-        values={{ BuildContext: <code>BuildContext</code> }}
-      >
-        {`Create/share/tests providers, with no dependency on Flutter. This
-        includes being able to listen to providers without a
-        {BuildContext}.`}
-      </Translate>
-    ),
-  },
-];
-
-function Feature({ imageUrl, title, description }) {
-  const imgUrl = useBaseUrl(imageUrl);
-  return (
-    <div className={classnames("col col--4", styles.feature)}>
-      {imgUrl && (
-        <div className="text--center">
-          <img className={styles.featureImage} src={imgUrl} alt={title} />
-        </div>
-      )}
-      <h3>{title}</h3>
-      <p>{description}</p>
-    </div>
-  );
-}
-
-function Preview({ imageUrl, title }) {
-  const imgUrl = useBaseUrl(imageUrl);
-  return (
-    <div className="col">
-      <h2>{title}</h2>
-      <img alt={title} src={imgUrl}></img>
-    </div>
-  );
-}
-
-function Home() {
-  const context = useDocusaurusContext();
-  const { siteConfig = {} } = context;
+export default function Home() {
+  const { siteConfig } = useDocusaurusContext();
   return (
     <Layout
       title={siteConfig.tagline}
@@ -310,5 +238,3 @@ function Home() {
     </Layout>
   );
 }
-
-export default Home;
