+++ website/src/pages/styles.module.scss
@@ -1,10 +1,3 @@
-/* stylelint-disable docusaurus/copyright-header */
-
-/**
- * CSS files with the .module.css suffix will be treated as CSS modules
- * and scoped locally.
- */
-
 .heroBanner {
   padding: 4rem 0;
   text-align: center;
@@ -31,17 +24,6 @@
   display: flex;
   align-items: center;
   justify-content: center;
-}
-
-.features {
-  align-items: center;
-  padding: 2rem 0;
-  width: 100%;
-}
-
-.featureImage {
-  height: 200px;
-  width: 200px;
 }
 
 .detailedFeatures {
